+++ packages/core/src/ambient.d.ts
@@ -22,14 +22,3 @@ declare module 'avsc/etc/browser/avsc.js' {
   const avsc: any;
   export = avsc;
 }
-
-declare module '@ljheee/xmind-parser' {
-  export function parseXmind8Xml(
-    xmlString: string,
-    options?: Record<string, unknown>
-  ): Promise<any[]>;
-  export function parseXmind2020Json(
-    contentJson: string | unknown[],
-    options?: Record<string, unknown>
-  ): any[];
-}